[rename] 이미지 파일 경로 수정 - 관련성을 기반으로 폴더 세분화 - 폴더 세분화에 따른 이미지 파일 경로 수정
