+++ src/data/category.js
@@ -1,8 +1,8 @@
 export const category = [
-  {img: '/total.svg', title: '전체', path: 'total'},
-  {img: '/recommend.svg', title: '추천', path: 'recommend'},
-  {img: '/vegetable.svg', title: '채소', path: 'vegetable'},
-  {img: '/fruit.svg', title: '과일', path: 'fruit'},
-  {img: '/grains.svg', title: '곡류', path: 'grains'},
-  {img: '/meat.svg', title: '육류', path: 'meat'},
+  {img: '/category/total.svg', title: '전체', path: 'total'},
+  {img: '/category/recommend.svg', title: '추천', path: 'recommend'},
+  {img: '/category/vegetable.svg', title: '채소', path: 'vegetable'},
+  {img: '/category/fruit.svg', title: '과일', path: 'fruit'},
+  {img: '/category/grains.svg', title: '곡류', path: 'grains'},
+  {img: '/category/meat.svg', title: '육류', path: 'meat'},
 ];
